FlashNotificationService: Trigger the camera flash for incoming notifications
This was originally designed as an accessibility setting for my custom ROM but
was integrated to work with a feature called Active Notifications.  It mimics
the camera flash for messaging feature in iOS and looks better than the apps
i've found in the playstore that only strobe. It's designed to work with the
torch.apk from cyanogenmod but can easily be adapted to use other LED interfaces.